Tell Jest to expect install test to fail if using an old version of npm
For npm@6 and older you have to use a different option to tell it to not
install optional dependencies. Unfortunately, adding that option to our
npmrc results in warnings for npm@8.6 and above. So instead we just
don't support npm@6 (see issue #1354). This means the tests will fail on
Node 12 and 14 (which come with npm@6 by default), so we mark the test
as expected to fail when npm is old.