Tag test release archives with the id of the stash commit
I found that when I was running integration tests locally it wasn't
regenerating the test release archive to pick up changes I had made; I
relised this was because it was using the HEAD commit ID as a name/tag,
which wasn't changing because I hadn't committed the changes I had made.
This isn't the desired behaviour, it should regenerate if changes have
been made to the working directory.

This commit reworks the logic of the test utils to make it so that the
git stash commit ID (if present) is used to name the test release
archive.

Note: git stash create is not deterministic, if you run it twice it will
give you two different commit IDs even if you haven't changed the files
in the working directory. I'm not sure why this is, or how to
workaround, but we probably shouldn't rely on git for creating release
archives for our tests for much longer unless we can find a fix.