Allow updating test prototype in tmp folder
The run-from-release script used to silently ignore a newer release
archive if there was already a test prototype in the directory; which in
hindsight isn't very helpful. Instead we let `mkPrototypeSync` overwrite
existing files. There's still a chance that I might get my test
prototype in a weird state, but there's a better chance now that it will
actually run the code I just wrote.