Add workaround in tests for buggy CSS updating
There is a bug where if you add a new stylesheet while on a page, the
page will blink to indicate that Browsersync has picked up a change, but
the styles won't actually change to reflect the new stylesheet.

@BenSurgisonGDS first discovered this bug when testing with Glitch, but
when working on these tests I found I could reproduce locally (on
macOS). I also checked whether this was a recent regression, but it
appears to have been around since at least version 12.0.3.

We should fix this issue, but not in this PR, so instead we'll add a
workaround and add a note to remove the workaround later. I've create
ticket #1440 to address this issue.