Add workaround in test for buggy image updating on Windows
For some reason the test works on macOS and Linux but not Windows; on
Windows we have to reload the page manaully for the image to show up.
This seems like a bug in our setup, but for now we just work around it.
We use waiting for resource instead of waiting for page load, and then
reload the page when the resource is present.

I've create issue #1440 to address this.