Refactor update script
We want to be able to test the code that handles command line
invocation, so we need to split that from the code that actually does
the thing. This commit splits the code to create the release archive
into its own function, so that we can also use the function in other
scripts in future.