Split create-release-archive code into sepearate modules
Unfortunately with Node.JS we can't mock functions within the module
under test, so this commit splits the create-release-archive script into
several different modules, allowing for clearer test boundaries. We put
the modules into a new folder `internal_lib`, to make it clear that this
code should not be included in releases and is only for use by tooling
and tests.