Add update script to update package.json for user
This PR adds an update script that compares the users package.json file
with the file we would expect them to have, and copies any changes we
think they made to add extra packages over to the upgraded package.json.
The desired effect is that after running the update script they should
not need to re-add any packages they personally installed previously.