Add internal docs on releasing from a support branch
We are planning to do a support release of the prototype kit soon.

We can base our process to do this on what the Design System team do for
GOV.UK Frontend [[1]]. This commit copies their documentation to our
repo, with a few changes as necessary to make it fit the prototype kit.

[1]: https://github.com/alphagov/govuk-frontend/blob/main/docs/releasing/publishing-from-a-support-branch.md