Fix imports of app/config
When using the kit as a package, the file app/config.js may not be
present. This commit makes it easier for code in lib to import
app/config.js by adding a convenience module lib/config.js, that imports
the user's file if it exists, and gives an empty object if it doesn't.

Also, when installing the kit as a package the app folder we can't rely
on relative paths to find the app folder, so we also add an appDir
property to the path-utils module to make it easier to get the appDir
path in code in lib.