Move copy of CHANGELOG.md to each project
We want each project in this repo to have a separate changelog, but we
don't need one in the root folder. To keep the history of changes around
we add an copy of the existing CHANGELOG.md to each project, but in
future they will have separate entries that look very different;
prototype-starter will include mainly changes to package versions and
occasionally changes to package templates, while the package project
will continue closer to as before.