Install govuk-prototype-kit package in a hacky way for now
To ease local development we install the package as a local project.
This isn't ideal because it's not how we want to release things, but
it's fine for now. Also we need to install express because for some
reason when using npm link we can't find express. It can find express
when installing from a tarball, but we probably should be re-exporting
or something explicit rather than relying on this weird behaviour. Also
it would be nice to use npm workspaces, but currently the scripts need to
reach into the node_modules folder to run the kit. So when we fix those
things we can change this setup to use npm workspaces and feel happy.