Remove ability to create release archives of previous releases
With everything moving around and changing it's difficult to get the
create-release-archive script to support creating archives of previous
releases. Luckily, as we removed the update script and its tests in
commit 958d644, we no longer need that functionality, so let's delete
the code.

In future if we need to get previous release archives whal I'll suggest
instead is downlading them from GitHub (with a local cache).