Move cli to `/bin` folder
The `/scripts` folder doesn't seem to show up when installing the kit as
a package from GitHub, which isn't great because we want the cli script
to be installed by users. I think this is because we have the scripts
folder set as `export-ignore`, moving the to the bin folder may help.