Start running npm audit whenever we run npm install
We used to have `audit=false` in our `.npmrc` file because we wanted
users who installed the kit from our repo to not be bothered by a bunch
of warnings about out-of-date packages. However, we have since a)
updated all the packages and b) decided to stop giving users an archive
of our repo to install the kit, so we can reenable the audit output from
npm.