Copy .gitignore to prototype starter with kit cli
Partially reverts changes in 664ad5 (Make prototype kit cli work outside
kit). In that commit we had to stop trying to copy the .gitignore file
because it wasn't included in the npm package. However, with the
`package.json` `files` list we can make sure that `.gitignore` gets
copied, so we can keep the copy and avoid the effort of duplication.