Split install cli command into a two stage process
We want to make sure that the version that the version of the kit the
user has installed at the end of the install process is the same as the
one used to do the install. The easiest and best way to do this is to
have two steps, with the first bootstrapping the second with the latest
version of the kit package, similar to how create-react-app works [[1]].

This commit splits the process in two within the cli script.

[1]: https://github.com/facebook/create-react-app/blob/main/packages/create-react-app/createReactApp.js