Ensure that create script installs same version of starter files and kit
Splits the process of creating a new prototype into two with copying the
starter files done after the version of the kit to be used has been
installed. This makes sure the user doesn't end up with starter files
that require features not available to the installed kit version. This
is similar to what the create-react-app tool does [[1]].

[1]: https://github.com/facebook/create-react-app/blob/main/packages/create-react-app/createReactApp.js