Clean up nodemon watch ignores
In the past we wanted to ignore changes to any files in
the cypress folder. This made sense when the prototype was running in
the kit project folder, but now we support kit as a package only that
isn't possible, and the downside of this line is that the kit won't work
if the prototype path contains `cypress` in any part.

Also, we no longer create a public folder, so we no longer need to
ignore that either.