Let exec inherit stdio
It is sometimes useful to be able to `exec` a command and have the
output go to the terminal. This commit updates our exec to let us do
that. The `child_process.exec` function does not support this, however
spawn does, and we can emulate exec by using a shell (similar to what
the `execa` package or `child_process` itself does).