Add create-prototype-and-start script
Adds a script intended for use with npm scripts in package.json. It
sort of works like `start-server-and-test`, only much simpler: it
creates a prototype in a temporary folder using `mkPrototype`, then runs
a shell command in the temporary folder. By default it runs `npm run
start`, although by supplying a command line argument it can run any
command, which is what makes it more useful than a script that creates
and starts the prototype all within node. It can then be composed with
`start-server-and-test`, without the need to pass the test directory
around.

For instance, for a simple test:

        npx start-server-and-test 'node scripts/create-prototype-and-start' 3000 test:smoke

The npm script `start` will be run in the test directory
($TMPDIR/govuk-prototype-kit-test by default, but can be overridden with
KIT_TEST_DIR), and the test script (`test:smoke` in this case) will be
run in the kit project directory.