Add serve and dev scripts
We want to make sure users have NODE_ENV=production set when their
prototypes are hosted online, because that controls a lot of important
behaviour including the password control and HTTPS redirects. To make
this easier, this commit adds a command to the cli tool that sets
NODE_ENV and then does the build and run. We also add these commands to
the starter files package scripts.