Make serve default command
We want to make sure that whatever service a user might be running their
prototype on online, they use the production scripts and configuration.
Unfortunately, there isn't a standard way to detect whether a project is
running in production or not; not all services set NODE_ENV [[1]], not
all services will read a Procfile, etc. The only surefire way we can
think of is to make it so that running `npm start` in a prototype runs
the production code by default.

[1]: https://github.com/alphagov/govuk-prototype-kit/issues/1202