Add warning to users if it looks like they have run npm start wrongly
We want to catch what we suspect will be a common case, where the user
runs `npm start` having been used to that from before, and then ends up
with a prototype that isn't working and no idea why. In those
circumstances we should give the users a heads-up by outputting
something to the terminal. While we can't be certain of when this might
have happened, we can make a reasonable guess, and the downside of
logging to a hosting services console erroneously is negligible.