Change default session store to cookie store when in production mode
Change the default configuration of the kit so that when the kit is run
in production (`NODE_ENV=production`) the cookie session store is used
instead of the memory session store. The impact of this is that session
data will default to be preserved between server restarts, and the user
will not see a warning from `express-session` in their server logs every
server restart.

This commit keeps memory store as the default when running the kit
locally, and the configuration can still be overridden.