Shrinkwrap dependencies
We want to make try and protect users from supply chain attacks similar
to incidents we've seen in the past (specifically rc and ua-parser-js
hijacking [[1], [2]], colors sabotage [3]). We can use a shrinkwrap file
to do this [[4]], by distributing `npm-shrinkwrap.json` in our package
users who install the kit will install the same dependency tree as we
develop with, meaning we can audit dependency updates for them.

[1]: https://docs.google.com/document/d/1ULJ6rbvNRN2S4QXao0aklzIZNDAseJnQ3WSu_I4hnHQ/
[2]: https://docs.google.com/document/d/1bsvRltcbOkgGzDx8tYV_58Q7xHOgXeGuPGALGq7aoIg/
[3]: https://www.bleepingcomputer.com/news/security/dev-corrupts-npm-libs-colors-and-faker-breaking-thousands-of-apps/
[4]: https://docs.npmjs.com/cli/v8/configuring-npm/npm-shrinkwrap-json