Fix unbranded template
When we added `application.css` to the plugins system [[1]], we
accidentally broke the unbranded layout. There isn't an alternative
quick fix because of the way the unbranded layout hooks into and
basically duplicates `application.css`, so we're just going to back out
that change and revisit after v13.

This PR removes `application.css` from extensionsConfig, and adds it to
the stylesheets template include. Now you don't need to have
`application.css` included in every page.

[1]: https://github.com/alphagov/govuk-prototype-kit/blob/bf4474e0b05957465c91b87d76e055250c0fd06c/server.js#L56