Fix integration tests so they run in test folder rather than in repo
Previously integration tests were running the kit incorrectly, with
`projectDir` being equal to the kit repo folder.

This was result in some weird behaviours: the `.tmp` folder was being
created in the repo directory rather than the project directory, and
tests of govuk-frontend assets were testing the versions in the repo,
not the version installed in the project.