Ensure that migrate script installs same version of starter files and kit
Similar to 5868df4893c31008, splits the process of migrating an existing
prototype into two with copying the starter files done after the version
of the kit to be used has been installed. This makes sure the users
doesn't end up with a mismatch between migrations and features available
to the installed kit version.