Simplify nodemon configuration
We have been seeing some weird bugs with watching the `.env` file, which
this change fixes by simplifying how we use nodemon.

Apart from `.env`, the only files the user might change that we want to
restart the app automatically for are now in their `app` folder. We can
therefore simplify our code by restricting the watch directives to that
folder, and remove a bunch of ignore filters.

This should also result in nodemon needing to watch many fewer files,
which should hopefully mean more reliable performance generally, apart
from fixing the dotenv bug.

This might be a breaking change for users who have put files their
prototype files outside the app folder, but in the past we've tried to
encourage them not to do that so it shouldn't affect too many users.