Make acceptance tests more realistic
Changes Cypress tests so they run against a prototype that is setup
similar to how users will have the kit installed.

Because we're using `--version local` for tests, npm is installing a
folder, and by default will just create a symlink to that folder, rather
than installing the package and all dependencies [[1]]. This means the
`node_modules` folder is laid out differently to how it would be for
real users, which can cause issues (see ticket #1741 for example). We
can change the behaviour however by setting the `install-links` config
variable to true [[2]]. This commit adds a line to set this in the
Cypress server start script, meaning that the test prototype is
installed in a more representative fashion.

This also lets us simplify the create-starter-prototype script a bit;
because local folder packages are no longer nested, we no longer need to
run `npm install` before using the test plugins (extensions).

[1]: https://docs.npmjs.com/cli/v8/commands/npm-install
[2]: https://docs.npmjs.com/cli/v8/using-npm/config#install-links