Update migration script to handle app/views/layout.html
We've moved app/views/layout.html to app/views/layouts/main.html in the
prototype starter, but we don't yet want the migration script to change
layouts in existing prototypes, because that would require a lot of
extra code to handle updating every Nunjucks template. This does mean
though that we need to tweak the script a bit to be able to copy from
the starter files to the existing layout.