Add backup default layout
In cases where the user has deleted their default layout file, we still
want their prototype to work, so we add a fallback in the kit. Because
the Nunjucks environment has the location in `lib/nunjucks` at a lower
priority, this file will only be used if the user doesn't have a
similarly named file in `app/views/layout`.