Make user Nunjucks files highest priority
We want to make sure that when users can override Nunjucks templates in
the kit and plugins by creating a file with the same path/filename. This
commit moves the `app/views` folder where a user keeps their templates
to the top of the Nunjucks folders list, so that Nunjucks will look
there for templates first.