Remove dev dependencies from published package
Fixes issue #1799.

We use npm shrinkwrap to try and improve security for users, however the
shrinkwrap file includes dev dependencies, and a bug in npm [[1]] means
that users are getting the dev dependencies whenever they create a new
prototype.

This commit changes the publishing process so that these dev
dependencies are removed from package.json and npm-shrinkwrap.json,
using the tool [clean-publish].

[1]: https://github.com/npm/cli/issues/4323

[clean-publish]: https://github.com/shashkovdanil/clean-publish#readme