Add isTest property to config object
This commit refactors our use of the IS_INTEGRATION_TEST environment
variable. This lets us centralise the condition code, and make it more
complicated while being consistent. We use this to make it so that we
can also look at the value of NODE_ENV to determine whether we are in a
test. By default, Jest sets NODE_ENV to 'test', so now you can run tests
with jest without having to remember to set IS_INTEGRATION_TEST in the
test file or on the command line.