Stop express complaining if error occurs after res has been sent
In cases where an error occurs after the response has already been sent,
currently the kit will emit a confusing log trace, with one log
message from the original error, and a second log message complaining
that we have tried to set headers after already sending them to the
client. To stop the second message from appearing, we shouldn't call
`res.send()` in our error handler when headers have already been sent,
and instead delegate to the default error handler as recommended by the
express documentation [[1]]. The default error handler already logs to
the console, so we do that before calling `console.error` ourselves.

We also add a test for this, which calls `setImmediate(next, err)`
before doing `res.send()`. This simulates the situation where something
goes wrong in a background thread. The motivating example, something
going wrong in `express-session`, actually does this [[2]].

Note that the test does not assert anything about the error message

    Error [ERR_HTTP_HEADERS_SENT]: Cannot set headers after they are sent to the client

from issue #1807, even though this commit is designed to fix that. I was
unable to get this error to appear in the integration tests; I'm not
sure this is because of express supressing logging, or because the
process is run differently in tests. However, in acceptance tests and
real-world tests this commit does fix the issue.

[1]: https://expressjs.com/en/guide/error-handling.html#the-default-error-handler
[2]: https://github.com/expressjs/session/blob/1010fadc2f071ddf2add94235d72224cf65159c6/index.js#L337