Rename start scripts
The name `start:package` is a bit of a legacy holdover from when the kit
was not a package. We should give the start scripts names that better
reflect what their purpose today.

While we're at it, we also expose the `scripts/create-prototype-and-run`
script as `npm run start:dev`. This script has the benefit of
automatically linking the prototype kit package, so not only is it
quicker to get started, it also picks up changes to package code
automatically.