Rename code that refer to plugins
Previously we called plugins 'extensions'. Now that we have decided on
a new name, we should make sure our code uses the new name where
appropriate.

This commit renames all functions and variables that have 'extensions'
in the name so that they now have 'plugins' in the name. In some cases
where the change of name could break user code, we keep both names (and
leave a TODO to remind us to clean things up later).