Remove direct dependency on acorn
In PR #667 we added `acorn` as a direct dependency to fix complaints
from npm about peer dependencies, due to a bug in npm [[1]]. That bug
now appears to have been fixed; removing `acorn` from our `package.json`
does not result in any peer dependency warnings. Also, as we now ensure
that users do not install dev dependencies, we can be confident that
they will not need `acorn`, as it is only required by dev dependencies.

[1]: https://github.com/alphagov/govuk-prototype-kit/pull/666