Make installing plugin during Cypress tests more robust
Currently after running npm install we test whether the plugin has
finished being installed by looking for a file within the node_modules
folder of the test prototype. However, the kit server usually takes a
second or two to register that a plugin has been installed and restart
itself, whereas the file is usually present before that.

This commit rewrites the check to look at the manage prototype plugins
page, and which plugins are listed as installed.

Note that this does mean that the `installPlugins` function now won't
work when doing a prodtest; however we don't currently test any plugins
in our prod tests, so this is fine for now.