Move call of listenOnAvailablePort to dev-server code
This fixes #462, as the `process.exit(0)` is no longer done within the
`nodemon` process. It also means we're not running portScanner every time
the kit restarts, and we can remove the code saving the port number to a
file.

Note changing the port number in `app/config.json` will now require a
manual restart.