Stop using after hooks in Cypress where we can use before hook
Cypress best practice is not to use an `after` hook unless it is
strictly necessary [[1]]. This commit refactors all our acceptance tests
so that any state setup is done before, not after, the spec. As well as
making it easier to see what the state was for a failed test, it also
should hopefully mean we can spend less time waiting for the kit to
restart.

The downside is that it's more work to isolate each test from each
other, as you have to make sure you reset everything that could affect
the test in the before hook, even if you don't change that file in the
test itself.

[1]: https://docs.cypress.io/guides/references/best-practices#Using-after-or-afterEach-hooks