Associate plugin search input and label
The search input on the plugin list page does not have a programatically label because the label is not currently associated with the input.

Rather than using the separate `govukLabel` macro, pass the label options to the  `govukInput` macro and let it take care of associating the label for us.

We can also use the `govuk-label--s` modifier rather than a font weight override.