Validate `govuk-prototype-kit.config.json` using the current package
The `validate-plugin` command checks the structure of the `govuk-prototype-kit.config.json`.
Code changes in the branch the GitHub action runs may affect what a valid config looks like.
So the action should run against the code of the branch the action is running on,
rather than pull an arbitrary version from npm (which may be well outdated).