Prevent rendering of navigation when there are no links
The manage-prototype layout is used for the error pages as well,
with no links to render. This create a 1px border under the header,
which is very visible when the brand refresh is active.

It wasn't before because the border was the same width as the blue border
under the header.