Remove unused styling from manage prototype sass
A lot of sass was added in the early prototyping phase of the management pages which was missed on production review. This removes all classes in the manage prototype sass that aren't referenced elsewhere in the codebase.

Tested locally to double check and everything looks the same as on main.