Fix or silence Sass deprecation warnings
Silence the `import` deprecation as we want to keep using `@import` for now.

`mixed-decls` could be fixed in the code, by moving the properties
in `.govuk-prototype-kit-manage-prototype-task-list__items` before
any inclusion of `govuk-media-query` (one of them hidden inside
`govuk-responsive-margin`). This ensures the order in the properties
in the Sass code is the same as in the output CSS as Sass 2.0 will hoist
declarations happening after nested blocks like CSS nesting specifies.