Add `npm-shrinkwrap.json` file to published package
`npm-shrinkwrap.json` was added to avoid supply chain attacks for users
of the Prototype Kit:
https://github.com/alphagov/govuk-prototype-kit/pull/1654/commits/7b8dfdd011e5618f4cfa7210e3c2d99a73742a56

However it doesn't get automatically added to the package published on `npm`,
meaning npm installs the latest compatible version of each dependency,
leaving them vulnerable to supply chain attacks.

This adds the file explicitely to the files in the package,
ensuring the versions installed for users of the Prototype Kit are
the ones at the time of release.