Ensure `clean-publish` runs off the `main` branch
Previously the command was running from the branch created for the release,
while it should technically the same content as `main`, the command
proposed if `clean-publish` fails first resets the working copy
to the `main` branch, so it makes sense to do it before the first run
of `clean-publish` as well.

Hoping it'll solve an issue we encountered where the release script
tried to publish to npm a version that was already published.