Migrate to use `@inquirer/confirm`
Prompts have been removed from `inquirer` since 10.0.0,
so we need to migrate our use of `inquirer.prompt` to the standalone
`confirm` prompt offered by `@inquirer/prompt`.

It does simplify the code a little as the prompt call now returns
the users answers, without having to give each prompt a name and picking the value
in an `answers` object.

The replacement has happened like for like, leaving the check for whether
being [run in a "text terminal"](https://nodejs.org/api/tty.html#tty) when asking
for the usage data, but not if changing port. It doesn't add any new test either.